refactor: reservations -> reservationService 변경
