+++ src/main/java/roomescape/controller/RoomescapeApiController.java
@@ -9,22 +9,22 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.domain.Reservation;
-import roomescape.domain.Reservations;
 import roomescape.dto.ReservationRequest;
 import roomescape.dto.ReservationResponse;
+import roomescape.service.ReservationService;
 
 @RestController
 public class RoomescapeApiController {
 
-    private final Reservations reservations;
+    private final ReservationService reservationService;
 
-    public RoomescapeApiController(final Reservations reservations) {
-        this.reservations = reservations;
+    public RoomescapeApiController(final ReservationService reservationService) {
+        this.reservationService = reservationService;
     }
 
     @GetMapping("/reservations")
     public List<ReservationResponse> findAllReservations() {
-        return reservations.findAll()
+        return reservationService.findAll()
                 .stream()
                 .map(ReservationResponse::from)
                 .toList();
@@ -33,17 +33,16 @@ public class RoomescapeApiController {
     @PostMapping("/reservations")
     public ResponseEntity<ReservationResponse> addReservation(@RequestBody ReservationRequest request) {
         Reservation reservation = request.toReservation();
-        long savedId = reservations.addReservation(reservation);
+        long savedId = reservationService.addReservation(reservation);
         if (savedId > 0) {
             return ResponseEntity.ok(ReservationResponse.from(reservation.withId(savedId)));
         }
         return ResponseEntity.badRequest().build();
-
     }
 
     @DeleteMapping("/reservations/{id}")
     public ResponseEntity<Void> removeReservation(@PathVariable long id) {
-        boolean removed = reservations.removeReservationById(id);
+        boolean removed = reservationService.removeReservationById(id);
         if (removed) {
             return ResponseEntity.ok().build();
         }
