+++ src/test/java/roomescape/domain/ReservationServiceTest.java
@@ -8,11 +8,12 @@ import java.util.Arrays;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import roomescape.service.ReservationService;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-class ReservationsTest {
+class ReservationServiceTest {
 
     @Test
     @DisplayName("같은 날짜 및 시간 예약이 존재하지 않을 경우, 예약 정보를 저장한 다음 id를 리턴한다")
@@ -21,18 +22,18 @@ class ReservationsTest {
         LocalDate date = LocalDate.of(2025, 4, 16);
         LocalTime time = LocalTime.of(10, 10);
         List<Reservation> currentReservation = new ArrayList<>();
-        Reservations reservations = new Reservations(new InMemoryReservationDAO(currentReservation));
-        reservations.addReservation(new Reservation("notSameReservation",
+        ReservationService reservationService = new ReservationService(new InMemoryReservationDAO(currentReservation));
+        reservationService.addReservation(new Reservation("notSameReservation",
                 LocalDate.of(2025, 1, 1),
                 LocalTime.of(10, 10)));
 
         //when
         Reservation reservation = new Reservation("test", date, time);
-        long savedId = reservations.addReservation(reservation);
+        long savedId = reservationService.addReservation(reservation);
 
         //then
         assertAll(
-                () -> assertThat(reservations.findAll()).hasSize(2),
+                () -> assertThat(reservationService.findAll()).hasSize(2),
                 () -> assertThat(savedId).isEqualTo(2L)
         );
 
@@ -52,11 +53,11 @@ class ReservationsTest {
         LocalDate date = LocalDate.of(2025, 4, 16);
         LocalTime time = LocalTime.of(10, 10);
         List<Reservation> currentReservation = provideReservationWithSameDateTime(date, time);
-        Reservations reservations = new Reservations(new InMemoryReservationDAO(currentReservation));
+        ReservationService reservationService = new ReservationService(new InMemoryReservationDAO(currentReservation));
 
         //when & then
         Reservation duplicated = new Reservation("test", date, time);
-        long savedId = reservations.addReservation(duplicated);
+        long savedId = reservationService.addReservation(duplicated);
         assertThat(savedId).isEqualTo(-1);
     }
 
@@ -67,11 +68,11 @@ class ReservationsTest {
         LocalDate date = LocalDate.of(2025, 4, 16);
         LocalTime time = LocalTime.of(10, 10);
         List<Reservation> currentReservation = provideReservationWithSameDateTime(date, time);
-        Reservations reservations = new Reservations(new InMemoryReservationDAO(currentReservation));
+        ReservationService reservationService = new ReservationService(new InMemoryReservationDAO(currentReservation));
         long existedId = currentReservation.getFirst().getId();
 
         //when
-        boolean actual = reservations.removeReservationById(existedId);
+        boolean actual = reservationService.removeReservationById(existedId);
 
         //then
         assertThat(actual).isTrue();
@@ -81,11 +82,11 @@ class ReservationsTest {
     @DisplayName("존재하지 않는 예약을 삭제하려는 경우 false를 리턴한다")
     void removeNotExistReservationById() {
         //given
-        Reservations reservations = new Reservations(new InMemoryReservationDAO(new ArrayList<>()));
+        ReservationService reservationService = new ReservationService(new InMemoryReservationDAO(new ArrayList<>()));
         long notExistId = 1;
 
         //when
-        boolean actual = reservations.removeReservationById(notExistId);
+        boolean actual = reservationService.removeReservationById(notExistId);
 
         //then
         assertThat(actual).isFalse();
